requirements.txt
langchain==0.2.0
langchain-community==0.2.0
langchain-openai==0.1.7
langchain-huggingface==0.0.3
openai==1.30.1
faiss-cpu==1.8.0
streamlit==1.33.0
pydantic==2.7.4
pandas
numpy
tiktoken==0.7.0
python-dotenv==1.0.1
sentence-transformers
transformers
torch